refactor(auth): move cookie policy internals
