+++ src/notebooklm/_auth/__init__.py
@@ -1,0 +1,1 @@
+"""Private authentication implementation package."""
